refactor: Changed data-type of some variables.

+++ src/Model/Node.cs
@@ -13,15 +13,15 @@ namespace graphSNA.Model
         // Nonunique name of the user (e.g., "Ali")
         public string Name { get; set; }
         // Numerical features required for dynamic weight calculation
-        public double Activity { get; set; }        // Feature I (Aktiflik)
-        public double Interaction { get; set; }     // Feature II (Etkileşim)
-        public double ConnectionCount { get; set; } // Feature III (Bağlantı Sayısı)
+        public float Activity { get; set; }        // Feature I (Aktiflik)
+        public float Interaction { get; set; }     // Feature II (Etkileşim)
+        public long ConnectionCount { get; set; } // Feature III (Bağlantı Sayısı)
 
         public System.Drawing.Point Location { get; set; }
         public System.Drawing.Color Color { get; set; } = System.Drawing.Color.LightBlue;
 
         // Constructor
-        public Node(string id, string name, double activity, double interaction)
+        public Node(string id, string name, float activity, float interaction)
         {
             this.Id = id;
             this.Name = name;
